+++ app/utils/cache-server.ts
@@ -1,7 +1,6 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { LRUCache } from "lru-cache";
 
-const cache = new LRUCache<any, any>({
+const cache = new LRUCache<string, object>({
   // 最大缓存条目数
   max: 100,
   // 缓存过期时间（毫秒）
@@ -13,25 +12,21 @@ const cache = new LRUCache<any, any>({
  * @param key 缓存键
  * @returns 缓存值，若不存在则返回 undefined
  */
-export const getCache = <T>(key: string): T | undefined => cache.get(key);
+export const getCache = <T extends object>(key: string): T | undefined =>
+  cache.get(key) as T | undefined;
 
 /**
  * 设置缓存
  * @param key 缓存键
  * @param value 缓存值
+ * @param ttl 可选的自定义过期时间（毫秒）
  */
-export const setCache = (key: string, value: any, ttl?: number): void => {
+export const setCache = (
+  key: string,
+  value: object | undefined,
+  ttl?: number,
+): void => {
+  if (!value) return;
   if (ttl) cache.set(key, value, { ttl });
   else cache.set(key, value);
 };
-
-/**
- * 删除缓存
- * @param key 缓存键
- */
-export const deleteCache = (key: string): boolean => cache.delete(key);
-
-/**
- * 清空所有缓存
- */
-export const clearCache = (): void => cache.clear();